Remove trailing whitespace and rename scancode.py as deps_info.py
